Implement name mangling for shared behaviors
This prevents tests (describe blocks, test functions) inside the block
that's importing a shared behavior from shadowing tests in the shared
behavior.  If we define a test with the same name in each block,
they are treated as two separate tests. Verbose output show the full
path to the test, including the name of the shared behavior.

Known issue: If multiple shared behaviors have the same name (e.g. they
are defined in separate modules, and imported into the test module with
different aliases), and have children with the same name, shadowing will
still occur.