Issue 32: Support list and tuple argnames for @pytest.mark.parametrize.
This commit affects the `pytestmark_name` method, which previously assumed that `mark.args[0]` was
always a str when converting the mark to a string.  However, `@pytest.mark.parametrize` permits an
`argnames` object which is a tuple or a list, for cases when you are parametrizing over multiple
variables.  This usage is now supported, and unit tests have been added.